+++ pubspec.yaml
@@ -3,25 +3,22 @@ description: "A new Flutter project."
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
-
 version: 1.0.0+1
 
 environment:
   sdk: ^3.7.2
 
-
 dependencies:
+  cupertino_icons: ^1.0.8
   flutter:
     sdk: flutter
-
- 
-  cupertino_icons: ^1.0.8
+  flutter_svg: ^2.2.1
+  gap: ^3.0.1
 
 dev_dependencies:
   flutter_lints: ^5.0.0
   flutter_test:
     sdk: flutter
-
 
 flutter:
   uses-material-design: true
